\name{plotrepresult}
\alias{plotrepresult}
\title{
Plot result of reconstruction.
}
\description{
Summary plot of the result of a network reconstruction.
}
\usage{
plotrepresult(lst, outfile, fontsize=15)
}
\arguments{
  \item{lst}{Return list of \code{\link{netga}}}
  \item{outfile}{Path to pdf file where to plot.}
  \item{fontsize}{Font size of text in the plot.}
}
\details{
TODO
}
\value{
TODO
}
\references{
TODO
}
\author{
Christian Bender
}
\note{
TODO
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
TODO
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ internal }

